lanenet-lane-detection/requirements.txt
loguru
tqdm
matplotlib
opencv_contrib_python
numpy
scikit_learn
PyYAML
tensorflow_gpu==1.15.0

